requirements.txt
langchain
langchain-openai
langchain-community
deepagents
python-dotenv
streamlit
pypdf
python-docx
pandas
langchain_groq